reports-service/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pandas==2.2.2
python-dotenv==1.0.1
pymysql==1.1.1
sqlalchemy==2.0.36
openpyxl==3.1.2
reportlab==4.2.0
